requirements.txt
torch
numpy
Pillow
tqdm
PyYAML
Flask
